finish crud view and api NewsArticle and viewHistory

+++ Group7_SE1733_A01_BE/Repositories/CategoryRepository.cs
@@ -17,7 +17,7 @@ namespace Repositories
             return await _context.Categories.ToListAsync();
         }
 
-        public async Task<Category> GetByIdAsync(int id)
+        public async Task<Category> GetByIdAsync(short id)
         {
             return await _context.Categories.FirstOrDefaultAsync(x => x.CategoryId == id);
         }
